refine the code
