+++ src/lib/components/AIReceiptImportDialog.svelte
@@ -1,12 +1,14 @@
 <script lang="ts">
-	import type { ExpenseRow, Currency, PreviewExpense, PreviewGroupExpense } from '$lib/types/expense';
+	import type {
+		ExpenseRow,
+		Currency,
+		PreviewExpense,
+		PreviewGroupExpense,
+		ReceiptResult
+	} from '$lib/types/expense';
 	import * as Dialog from '$lib/components/shadcn/dialog';
-	import { Button } from '$lib/components/shadcn/button';
+	// import { Button } from '$lib/components/shadcn/button';
 	import { user } from '$lib/stores/session.store';
-	import { PUBLIC_GOOGLE_AI_GCF, PUBLIC_SUPABASE_URL } from '$env/static/public';
-	import { supabase } from '$lib/supabase/supabaseClient';
-	import Logger from '$lib/utils/logger';
-	import * as jose from 'jose';
 	import { Sparkles, Upload, CircleCheckBig } from 'lucide-svelte';
 
 	// Child components
@@ -14,6 +16,7 @@
 	import UploadStep from './ai-receipt/UploadStep.svelte';
 	import AnalysisStep from './ai-receipt/AnalysisStep.svelte';
 	import ConfirmStep from './ai-receipt/ConfirmStep.svelte';
+	// import { mockReceiptResult } from './ai-receipt/mock';
 
 	let {
 		open = $bindable(false),
@@ -23,27 +26,14 @@
 		onImport?: (expenses: ExpenseRow[]) => void;
 	} = $props();
 
-	let aiStep = $state(3);
+	let aiStep = $state(1);
 	let aiUploading = $state(false);
 	let aiAnalyzing = $state(false);
 	let aiConverting = $state(false);
 	let selectedFile = $state<File | null>(null);
 	let previewUrl = $state<string | null>(null);
 	let lastUploadedFilePath = $state('');
-	// uploads/4b39017c-3ebb-466a-8ed2-3cbbfc8340e2/20260512032836_IMG_4146.jpg
 
-	interface ReceiptItem {
-		name: string;
-		quantity: number;
-		price: number;
-	}
-	interface ReceiptResult {
-		store_name: string;
-		date: string;
-		total_amount: number;
-		currency: string;
-		items: ReceiptItem[];
-	}
 	let analysisResult = $state<ReceiptResult | null>(null);
 
 	const wizardSteps = [
@@ -81,7 +71,7 @@
 	});
 
 	function resetAIDialog() {
-		// aiStep = 1;
+		aiStep = 1;
 		selectedFile = null;
 		analysisResult = null;
 		aiUploading = false;
@@ -104,89 +94,6 @@
 		}
 	});
 
-	async function callAIGCF(fileName: string, contentType: string) {
-		const {
-			data: { session }
-		} = await supabase.auth.getSession();
-		const accessToken = session?.access_token;
-		if (!accessToken) {
-			console.error('No access token found');
-			return;
-		}
-
-		try {
-			const JWKS = jose.createRemoteJWKSet(
-				new URL(`${PUBLIC_SUPABASE_URL}/auth/v1/.well-known/jwks.json`)
-			);
-			const { payload } = await jose.jwtVerify(accessToken, JWKS);
-			Logger.log('Access token verified using jose:', payload);
-		} catch (error) {
-			console.error('Token verification failed:', error);
-			return;
-		}
-
-		try {
-			const url = new URL(PUBLIC_GOOGLE_AI_GCF);
-			url.searchParams.set('action', 'get_upload_url');
-			url.searchParams.set('file_name', fileName);
-			url.searchParams.set('content_type', contentType);
-
-			const response = await fetch(url.toString(), {
-				headers: {
-					Authorization: `Bearer ${accessToken}`
-				}
-			});
-			const signedUrl = await response.json();
-			Logger.log('Received signed URL from GCF:', signedUrl);
-			return signedUrl;
-		} catch (error) {
-			console.error('Error calling AI GCF:', error);
-			return null;
-		}
-	}
-
-	async function callAnalyzeReceipt(filePath: string) {
-		if (!filePath) {
-			return;
-		}
-		aiStep = 2;
-		aiAnalyzing = true;
-		analysisResult = null;
-
-		const {
-			data: { session }
-		} = await supabase.auth.getSession();
-		const accessToken = session?.access_token;
-		if (!accessToken) {
-			console.error('No access token found');
-			aiAnalyzing = false;
-			return;
-		}
-
-		try {
-			const url = new URL(PUBLIC_GOOGLE_AI_GCF);
-			url.searchParams.set('action', 'analyze_receipt');
-			url.searchParams.set('file_path', filePath);
-
-			Logger.log('Triggering AI analysis for:', filePath);
-			const response = await fetch(url.toString(), {
-				headers: {
-					Authorization: `Bearer ${accessToken}`
-				}
-			});
-
-			const data = await response.json();
-			if (data.status === 'success' && data.result) {
-				analysisResult = data.result;
-			}
-			Logger.log('AI Analysis Result:', data);
-		} catch (error) {
-			console.error('Error calling analyze_receipt:', error);
-		} finally {
-			aiAnalyzing = false;
-		}
-	}
-
 	function handleConfirm(expenses: ExpenseRow[]) {
 		onImport?.(expenses);
 		open = false;
@@ -202,51 +109,15 @@
 			{aiAnalyzing}
 			onBack={() => (aiStep -= 1)}
 		/>
-		<Button
+		<!-- <Button
 			variant="outline"
 			onclick={() => {
-				analysisResult = {
-					store_name: '好市多股份有限公司汐止分公司',
-					date: '2026-04-18',
-					total_amount: 2708,
-					currency: 'TWD',
-					items: [
-						{
-							name: '桂格有機燕麥片',
-							quantity: 1,
-							price: 415
-						},
-						{
-							name: 'IRISWOOZOO循環扇',
-							quantity: 1,
-							price: 1229
-						},
-						{
-							name: '蚵仔煎洋芋片',
-							quantity: 1,
-							price: 198
-						},
-						{
-							name: '北海鱈魚香絲600G',
-							quantity: 1,
-							price: 299
-						},
-						{
-							name: '富士山氣泡水檸檬',
-							quantity: 1,
-							price: 369
-						},
-						{
-							name: '黑標紅酒瓶售',
-							quantity: 1,
-							price: 198
-						}
-					]
-				};
+				analysisResult = mockReceiptResult;
+				aiStep = 3;
 			}}
 		>
-			Load data
-		</Button>
+			Load mock data
+		</Button> -->
 
 		<div class="flex-1 overflow-y-auto px-1">
 			{#if aiStep === 1}
@@ -254,22 +125,20 @@
 					bind:aiStep
 					bind:aiUploading
 					bind:aiConverting
-					{aiAnalyzing}
+					bind:aiAnalyzing
 					bind:selectedFile
 					{previewUrl}
 					bind:lastUploadedFilePath
-					{callAIGCF}
-					{callAnalyzeReceipt}
+					bind:analysisResult
 				/>
 			{:else if aiStep === 2}
 				<AnalysisStep
 					bind:aiStep
 					{aiUploading}
-					{aiAnalyzing}
-					{analysisResult}
+					bind:aiAnalyzing
+					bind:analysisResult
 					{previewUrl}
 					{lastUploadedFilePath}
-					{callAnalyzeReceipt}
 					onReset={() => (aiStep = 1)}
 				/>
 			{:else if aiStep === 3}
